+++ craft_application/commands/lifecycle.py
@@ -22,16 +22,13 @@ import subprocess
 import textwrap
 from typing import Any, Literal, cast
 
-import pydantic
 from craft_cli import CommandGroup, CraftError, emit
 from craft_parts.features import Features
 from typing_extensions import override
 
 from craft_application import errors, util
 from craft_application.commands import base
-from craft_application.errors import TestFileError
 from craft_application.util import ProServices
-from craft_application.util.error_formatting import format_pydantic_errors
 from craft_application.util.logging import handle_runtime_error
 
 
@@ -652,10 +649,9 @@ class PackCommand(LifecycleCommand):
 class TestCommand(PackCommand):
     """Command to run project tests.
 
-    The test command invokes the spread command with a processed spread.yaml
-    configuration file.
+    The test command invokes spread with a processed test config file.
 
-    This command is opt-in for applications in craft-application 5 and will become
+    This command is opt-in for applications and will become
     a standard lifecycle command in a future major release.
     """
 
@@ -695,15 +691,8 @@ class TestCommand(PackCommand):
         parsed_args.output = pathlib.Path.cwd()
 
         testing_service = self._services.get("testing")
-        # Fail early if spread.yaml is invalid.
-        try:
-            testing_service.parse_spread_yaml()
-        except pydantic.ValidationError as exc:
-            raise TestFileError(
-                format_pydantic_errors(exc.errors(), file_name="spread.yaml"),
-                reportable=False,
-                retcode=os.EX_DATAERR,
-            )
+        # Fail early if the test config file is invalid.
+        testing_service.parse_spread_yaml()
 
         if util.is_managed_mode():
             # If we're in managed mode, we just need to pack.